FFB effects reversed
The force feedback effects in steering wheel mode(which emulates the Logitech SpeedForce Wheel) for the GameCube controller plugin are reversed, even when setting the slider to -100 in the Rumble Motor menu. This is with all effects, *including* Constant. Developer SkidAu made a change in the GameCube pad plugin to fix this very same issue about three years ago, but looks like those changes have been reverted in one of the branches following his fix.